hilder_auction/requirements.txt
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
idna==2.7
lxml==4.2.5
lxmlxtree==0.0.3
py==1.7.0
pymongo==3.7.1
PyYAML==3.13
qiniu==7.2.2
requests==2.19.1
retry==0.9.2
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
urllib3==1.23
yaml-1.3==0.1.0
